requirements.txt
packaging==24.1
pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
